+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support of subdirectories in nextcloud webdav url. Fixes the problem with tagging methods return a 404 
+### Fixed
+- fixed perfromace problems getFile and getFolder
 
 ## 0.1.3 - 2019-01-09
 ### Added